mod for models and controller

+++ app/Http/Controllers/BioController.php
@@ -4,7 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Bio;
 use App\Models\zat;
-use App\Models\dataSenyawa;
+use App\Models\sbt;
+use App\Models\tanaman;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
@@ -18,7 +19,7 @@ class BioController extends Controller
     public function index()
     {
         $Bio = Bio::get();
-        return view('clientData.listBio', ['Bio' => $Bio]);
+        return view('clientData.listBio', compact('Bio'));
     }
 
     /**
@@ -28,9 +29,10 @@ class BioController extends Controller
      */
     public function create()
     {
-        $dataSenyawa = dataSenyawa::get();
+        $tanaman = tanaman::get();
         $zat = zat::get();
-        return view('halaman.inputData3', ['dataSenyawa' => $dataSenyawa], ['zat' => $zat]);
+        $sbt = sbt::get();
+        return view('halaman.inputData3', compact('tanaman', 'zat', 'sbt'));
     }
 
     /**
@@ -52,13 +54,29 @@ class BioController extends Controller
         $plantNames = $request->input('Plant_Name');
         $fitokimia = $request->input('Phytochemical');
 
-        DB::table('Bio')->insert([
+        $baId = DB::table('bioaktivitas')->insertGetId([
             'BA_Name' => $request['BA_Name'],
             'BA_Details' => $request['BA_Details'],
             'BA_ref' => $request['BA_ref'],
-            'Plant_Name' => json_encode($plantNames),
-            'Phytochemical' => json_encode($fitokimia),
         ]);
+
+        if (is_array($plantNames) && is_array($fitokimia)) {
+            foreach ($plantNames as $plantName) {
+                foreach ($fitokimia as $Fitokimia) {
+                    DB::table('sbt')->insert([
+                        'snywId' => $Fitokimia,
+                        'tanId' => $plantName,
+                        'biokId' => $baId, 
+                    ]);
+                }
+            }
+        } else {
+            DB::table('sbt')->insert([
+                'snywId' => is_array($fitokimia) ? $fitokimia[0] : $fitokimia,
+                'tanId' => is_array($plantNames) ? $plantNames[0] : $plantNames,
+                'biokId' => $baId,
+            ]);
+        }
     
         return redirect('/Bio');
     }
@@ -71,11 +89,18 @@ class BioController extends Controller
      */
     public function show($id)
     {
-        $Bio = Bio::all()->where('id', $id)->first();
-        $zat = zat::get();
-        $dataSenyawa = dataSenyawa::get();
+        $Bio = Bio::findOrFail($id);
+
+        $sbtItems = Sbt::where('biokId', $Bio->id)->get();
+
+        $tanIds = $sbtItems->pluck('tanId');
+        $snywIds = $sbtItems->pluck('snywId');
+
+        $tanNames = tanaman::whereIn('id', $tanIds)->pluck('Plant_Name', 'id');
+
+        $senyawaNames = zat::whereIn('id', $snywIds)->pluck('Phytochemical', 'id');
         
-        return view('detail.dataDetail3', compact('Bio', 'zat', 'dataSenyawa'));
+        return view('detail.dataDetail3', compact('tanNames' , 'senyawaNames', 'Bio'));
     }
 
     /**
@@ -86,11 +111,18 @@ class BioController extends Controller
      */
     public function edit($id)
     {
-        $Bio = Bio::all()->where('id', $id)->first();
-        $zat = zat::get();
-        $dataSenyawa = dataSenyawa::get();
+        $Bio = Bio::findOrFail($id);
+
+        $sbtItems = Sbt::where('biokId', $Bio->id)->get();
+
+        $tanIds = $sbtItems->pluck('tanId');
+        $snywIds = $sbtItems->pluck('snywId');
+
+        $tanNames = tanaman::whereIn('id', $tanIds)->pluck('Plant_Name', 'id');
+
+        $senyawaNames = zat::whereIn('id', $snywIds)->pluck('Phytochemical', 'id');
         
-        return view('update.updateData3', compact('Bio', 'zat', 'dataSenyawa'));
+        return view('update.updateData3', compact('tanNames' , 'senyawaNames', 'Bio'));
     }
 
     /**
@@ -113,17 +145,33 @@ class BioController extends Controller
         $plantNames = $request->input('Plant_Name');
         $fitokimia = $request->input('Phytochemical');
 
-        $Bio = Bio::all()
-        ->where('id', $id)
-        ->first();
+        $Bio = Bio::find($id);
         if ($Bio) {
             $Bio->update([
             'BA_Name' => $request['BA_Name'],
             'BA_Details' => $request['BA_Details'],
             'BA_ref' => $request['BA_ref'],
-            'Plant_Name' => json_encode($plantNames),
-            'Phytochemical' => json_encode($fitokimia),
         ]);}
+
+        DB::table('sbt')->where('biokId', $id)->delete();
+
+        if (is_array($plantNames) && is_array($fitokimia)) {
+            foreach ($plantNames as $plantName) {
+                foreach ($fitokimia as $Fitokimia) {
+                    DB::table('sbt')->insert([
+                        'snywId' => $Fitokimia,
+                        'tanId' => $plantName,
+                        'biokId' => $id, 
+                    ]);
+                }
+            }
+        } else {
+            DB::table('sbt')->insert([
+                'snywId' => is_array($fitokimia) ? $fitokimia[0] : $fitokimia,
+                'tanId' => is_array($plantNames) ? $plantNames[0] : $plantNames,
+                'biokId' => $id,
+            ]);
+        }
     
         return redirect('/Bio');
     }
@@ -136,8 +184,12 @@ class BioController extends Controller
      */
     public function destroy($id)
     {
-        $Bio = Bio::where('_id', $id)->first();
-        $Bio->delete();
-        return redirect('/Bio')->with('success','Data Berhasil Dihapus!');
+        $Bio = Bio::find($id);
+        if ($Bio) {
+            $Bio->delete();
+            return redirect('/Bio')->with('success', 'Data Berhasil Dihapus!');
+        } else {
+            return redirect('/Bio')->with('error', 'Data Tidak Ditemukan!');
+        }
     }
 }
